Rename and expose ResultListener as  TextWriterResultChannel
