+++ src/xunit.runner.devices/Platforms/UWP/UI/RunnerApplication.cs
@@ -10,6 +10,7 @@ using Windows.UI.Core;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Navigation;
+using Xunit.Runners.ResultChannels;
 
 namespace Xunit.Runners.UI
 {
@@ -18,11 +19,9 @@ namespace Xunit.Runners.UI
         readonly List<Assembly> testAssemblies = new List<Assembly>();
 
         public bool TerminateAfterExecution { get; set; }
-
         [Obsolete("Use ResultChannel")]
-        protected TextWriter Writer { get; set; }
-        protected IResultChannel ResultChannel { get; set; }
-
+        public TextWriter Writer { get; set; }
+        public IResultChannel ResultChannel { get; set; }
         public bool AutoStart { get; set; }
         public bool Initialized { get; private set; }
 
@@ -81,7 +80,7 @@ namespace Xunit.Runners.UI
                     var nav = new Navigator(rootFrame);
 
 #pragma warning disable CS0618 // Type or member is obsolete
-                    var runner = new DeviceRunner(testAssemblies, nav, ResultChannel ?? new ResultListener(Writer));
+                    var runner = new DeviceRunner(testAssemblies, nav, ResultChannel ?? new TextWriterResultChannel(Writer));
 #pragma warning restore CS0618 // Type or member is obsolete
                     var hvm = new HomeViewModel(nav, runner);
