+++ src/xunit.runner.devices/Platforms/android/TextWriterResultChannel.cs
@@ -11,9 +11,9 @@ using Android.OS;
 using Android.Widget;
 using Xunit.Runners.UI;
 
-namespace Xunit.Runners
+namespace Xunit.Runners.ResultChannels
 {
-    class ResultListener : IResultChannel
+    public class TextWriterResultChannel : IResultChannel
     {
         int failed;
         int passed;
@@ -21,7 +21,7 @@ namespace Xunit.Runners
         TextWriter writer;
         readonly object lockOjb = new object();
 
-        public ResultListener(TextWriter writer)
+        public TextWriterResultChannel(TextWriter writer)
         {
             this.writer = writer;
         }
