+++ src/xunit.runner.devices/ResultChannels/Extensions.cs
