+++ src/xunit.runner.devices/ResultChannels/FilteredCollectionView.cs
