+++ src/xunit.runner.devices/ResultChannels/IResultChannel.cs
